Moved `components/` and `pages/` to `ui/`
