+++ frontend/src/ui/components/shared/index.ts
@@ -1,3 +1,3 @@
-export * from './ui-elements';
+export * from './misc';
 export * from './navigation';
 export * from './form-elements';
