+++ frontend/src/ui/components/shared/misc/avatar.tsx
