+++ frontend/src/ui/components/shared/misc/modal.tsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom';
 import { CSSTransition } from 'react-transition-group';
 
 import { Backdrop } from '..';
-import { OnClickEvent } from '../../../util';
+import { OnClickEvent } from '../../../../util';
 
 
 interface ModalOverlayProps {
